Clean-up refactoring: moved "ocaml.editor.newFormatter" package into "ocaml.editor.formatter" package.
