+++ Ocaml/src/ocaml/editor/formatting/Pos.java
@@ -1,4 +1,4 @@
-package ocaml.editor.newFormatter;
+package ocaml.editor.formatting;
 
 import beaver.Symbol;
