+++ Ocaml/src/ocaml/editor/formatting/Test.java
@@ -1,4 +1,4 @@
-package ocaml.editor.newFormatter;
+package ocaml.editor.formatting;
 
 import java.io.IOException;
 import java.io.StringReader;
